Added parens parsing
Predicting the tree structure in my own mind became to cumbersome, so
added ast_postfix method to allow me to serialize it as a string, and
began writing tests accordingly.

Abstracted operator_t out to its own file to represent the fact that
it is its own concept independent of the parser.

The parser is a bit of an ugly mess at the moment because of the
constant need to keep looking up operators for precedence.